New MultiServerAutocompleteInput component
