+++ src/frontend/packages/input-components/package.json
@@ -18,7 +18,7 @@
   },
   "peerDependencies": {
     "react": "^16.13.1 || ^17.0.0",
-    "react-admin": "^3.13.1"
+    "react-admin": "^3.19.10"
   },
   "publishConfig": {
     "access": "public"
