+++ src/frontend/packages/input-components/src/OptionRenderer.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+const OptionRenderer = ({ record, optionText, dataServers }) => {
+  const server = Object.values(dataServers).find(server => record.id.startsWith(server.baseUrl));
+  return (
+    <span>
+      {record[optionText]}
+      {server && (
+        <em className="serverName" style={{ color: 'grey' }}>
+          &nbsp;({server.name})
+        </em>
+      )}
+    </span>
+  );
+};
+
+export default OptionRenderer;
