+++ src/frontend/packages/input-components/src/index.js
@@ -1,1 +1,2 @@
 export { default as MultiServerAutocompleteArrayInput } from './MultiServerAutocompleteArrayInput';
+export { default as MultiServerAutocompleteInput } from './MultiServerAutocompleteInput';
